fix: 3 critical bugs before first live session
1. Bracket fallback: when Alpaca rejects bracket order (wash-trade check,
   API error, version issue), signal was silently dropped. Now falls back
   to plain market entry so no trade is missed.

2. Fractional qty truncation: get_positions() used int(p.qty) which
   truncates 0.5 → 0. Reconcile then saw a "ghost position" and called
   close_position() on a real open Alpaca position, stopping all stop/
   trailing management on a live trade. Fixed: float(p.qty) everywhere.

3. Double stop on bracket partial exit: T1 partial exit called both
   place_stop_order() and modify_stop_loss() for bracket positions,
   creating two simultaneous stop orders with a 1-second race window
   that could double-sell and open an unintended short. Bracket positions
   (is_bracket=True on Position) now skip broker-side stop manipulation —
   Alpaca manages its child SL internally.

Also: added is_bracket field to Position dataclass.

+++ data_fetch_alpaca.py
@@ -738,7 +738,7 @@ class AlpacaDataFetcher:
             for p in positions:
                 result.append({
                     "symbol":     p.symbol,
-                    "qty":        int(p.qty),
+                    "qty":        float(p.qty),  # float: fractional shares have qty like 0.5
                     "side":       p.side.value,   # "long" or "short"
                     "avg_price":  float(p.avg_entry_price),
                     "market_val": float(p.market_value or 0),
